+++ backend/accounts/admin.py
@@ -1,0 +1,31 @@
+from django.contrib import admin
+from django.contrib.auth.models import Group
+from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
+
+from .models import User
+from .admin_forms import CustomUserChangeForm, CustomUserCreationForm
+
+class UserAdmin(BaseUserAdmin):
+    form = CustomUserChangeForm
+    add_form = CustomUserCreationForm
+
+    list_display = ("email", "name", "is_staff")
+    list_filter = ("is_staff",)
+    fieldsets = (
+        (None, {"fields": ("email", "password")}),
+        ("Personal info", {"fields": ("name",)}),
+        ("Permissions", {"fields": ("is_staff", "is_active")}),
+    )
+    add_fieldsets = (
+        (None, {"classes": ("wide",), "fields": ("email", "password1", "password2")}),
+    )
+    search_fields = ("email",)
+    ordering = ("email",)
+    filter_horizontal = ()
+
+admin.site.register(User, UserAdmin)
+admin.site.site_header = "Vote App Admin"
+admin.site.site_title = "Vote App Admin Portal"
+admin.site.index_title = "Welcome to Vote App Admin Portal"
+admin.site.site_url = "http://127.0.0.1:8000/"
+admin.site.unregister(Group)